+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v0.27.2] - 2022-11-06
+
 - mark alternate register accessors with `const`, bump `pac` MSRV to 1.61
 - `fields` fn refactoring
 - Test patched STM32
@@ -772,7 +774,8 @@ peripheral.register.write(|w| w.field().set());
 
 - Initial version of the `svd2rust` tool
 
-[Unreleased]: https://github.com/rust-embedded/svd2rust/compare/v0.27.1...HEAD
+[Unreleased]: https://github.com/rust-embedded/svd2rust/compare/v0.27.2...HEAD
+[v0.27.2]: https://github.com/rust-embedded/svd2rust/compare/v0.27.1...v0.27.2
 [v0.27.1]: https://github.com/rust-embedded/svd2rust/compare/v0.27.0...v0.27.1
 [v0.27.0]: https://github.com/rust-embedded/svd2rust/compare/v0.26.0...v0.27.0
 [v0.26.0]: https://github.com/rust-embedded/svd2rust/compare/v0.25.1...v0.26.0